[14.0][FIX] base_report_to_printer: Close dialog on direct print.
Ensure wizard is closed if direct print is initiated from a wizard button.

The _triggerDownload function on base Odoo ActionManager explicitly closes
any open wizard dialogs in the UI.

This commit adds the same code to _triggerDownload for server print.

Without this code if a server print is initiated from a wizard button the
wizard dialog remains open in the UI, even though the print has completed
successfully.