[FIX] base_report_to_printer: do not download text report sent to printer
In the case of a report of type "qweb-text" set to "Send to printer" (e.g. a zpl report), the pervious implementation was triggering a download of the report.

This is because inside the base odoo report handling function (odoo/addons/web/static/src/webclient/actions/action_service.js -> `_executeReportAction`), the code defaults to 'download' for text reports that have not been dealt with and current implementation only deals with "qweb-pdf" reports.
However, the supported types defined inside `REPORT_TYPES` are ["qweb-pdf", "qweb-text"].

Thus, the current approach now creates a new computed field `is_report_to_printer_action` (to replace the previous hard-coded solution) to deal with the logic of ensuring the report is of the expected type before triggering our custom action handler.