FIX] base_report_to_printer: return a close action when the option is specified.
Many modules can return the key close_on_report_download in the context. In the web module, this option is expected and handled accordingly. However, this module ignored the option, causing the wizard not to close when a report is generated from an action with target 'new' (typically a wizard). After this commit, the wizard closes correctly.

https://github.com/odoo/odoo/blob/401540aaa8e42243f4504c594d477330295ddee0/addons/web/static/src/webclient/actions/action_service.js#L1326-L1330